+++ fail/fail.go
@@ -49,6 +49,7 @@ const (
 	ErrSomeDirsOnlyInTemplates = "@use, @insert, @reserve, @component only allowed in templates"
 	ErrGlobalFuncMissing       = "global function '%s' not found"
 	ErrProperyOnNonObject      = "cannot get propery '%s' from a non-object type '%s'"
+	ErrEachDirWithNonArrArg    = "cannot use @each statement with non-array type '%s' after 'in' keyword"
 
 	// Functions
 	ErrNoFuncForThisType  = "function '%s' not available for type '%s'"