+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 3. Fixed the `contains` function for arrays; `{{ ![{}, 21].contains({age: 21}) }}` now returns the correct result.
 4. Now you will get a proper error when trying to access a property on a non-object type like `{{ "str".nice }}`. Before, you would get a panic.
 5. Fixed issue where you couldn't write a slot directive with space after the `@slot` keyword. This `@slot ("book")` was giving an error previously.
+6. Now you will get an proper error when trying to use `@each` statement on non-array type.
 
 ### ✨ New Features
 1. Added `globals` object. You can now add `GlobalData` to your configurations and access this data in your templates using the `globals` object. For example: `globals.env`.